Add COBS encoding utility
Adds COBS encoding/decoding utilities. COBS will be the algorithm
used to unambigously frame packet data. COBS is an encoding scheme
that replaces any delimiter byte in a message with a non-delimiter
value. This allows us to use the delimiter byte (typically zero) as
a packet boundry. It also allows clients to easily re-sync to a
logging stream. They just need to listen for the next zero.